// src/interfaces/mkb/IMkb.ts
export default interface IMkb {
  id: string;
  class: string;
  className: string;
  classRange: string;
  classComment: string;
  groupName: string;
  groupRange: string;
  groupComment: string;
  subGroupName: string;
  subGroupRange: string;
  subGroupComment: string;
  subSubGroupName: string;
  subSubGroupRange: string;
  subSubGroupComment: string;
  diagnosisName: string;
  diagnosisComment: string;
  subDiagnosisName: string;
  subDiagnosisComment: string;
}